#7: Refactored
