+++ src/components/module/gacha/GachaDetail/GachaDetail.module.css
