+++ src/components/module/gacha/GachaDetail/GachaDetail.stories.tsx
