+++ src/components/module/gacha/GachaItem/GachaItem.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable react/require-default-props */
 import { FC, ReactNode } from 'react';
 import clsx from 'clsx';
-import { IconPlay } from '../../Icon';
-import Button from '../Button/Button';
+import { IconPlay } from '~/components/Icon';
+import Button from '~/components/ui/Button/Button';
 import styles from './GachaItem.module.css';
 import GachaPreview from './GachaPreview/GachaPreview';
 import { useRouter } from 'next/router';
