+++ src/components/module/gacha/GachaItem/GachaPreview/GachaPreview.tsx
