+++ src/components/module/gacha/PartyStage/PartyStage.module.css
