+++ src/components/module/gacha/PartyStage/PartyStage.stories.tsx
