+++ src/components/module/result/DisplayResult/DisplayResult.module.css
