+++ src/components/module/result/DisplayResult/DisplayResult.stories.tsx
