+++ src/components/module/result/DisplayResult/DisplayResult.tsx
