+++ src/components/module/result/DisplayResult/DisplayResultLine.tsx
